==> 04_rfmix/including_mais_samples/three_way/01.3_make-reference-sample-files.Rmd <==
---
title: '01.1_prepare-sample-files.Rmd'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Setup 

```{r}
"%&%" <- function(a,b) paste0(a,b)
library("data.table")
library("tidyverse")
serv.dir <- "./"
work.dir.serv <- serv.dir %&% 
  "popgen/04_rfmix/including_mais_samples/three_way/"
input.dir <- work.dir.serv %&% "input_files/"
output.dir <- work.dir.serv %&% "output_files/"
```

```{r}
samp.df <- fread(input.dir %&% "reference-sample-list.txt",header=F)
```

Read in pre-selected reference samples based on TeraStructure analysis 
```{r}
tera.dir <- serv.dir %&% "popgen/02.2_terastructure/" %&%
  "mais_amr_afr_eur_eas/output_files/"
afr.df <- fread(tera.dir %&% "samples-k18-p90-afr.txt",header=F)
eur.df <- fread(tera.dir %&% "samples-k18-p90-eur.txt",header=F)
amr.df <- fread(tera.dir %&% "samples-k18-p90-amr.txt",header=F)
```

Extract individuals 
```{r}
keep.afr <- purrr::map(afr.df$V1,function(s){
  samp.df$V1[grepl(pattern=s,x=samp.df$V1)]
}) %>% unlist(.) %>% as.character(.)  %>% unique(.) 
keep.eur <- purrr::map(eur.df$V1,function(s){
  samp.df$V1[grepl(pattern=s,x=samp.df$V1)]
}) %>% unlist(.) %>% as.character(.)  %>% unique(.) 
keep.amr <- purrr::map(amr.df$V1,function(s){
  samp.df$V1[grepl(pattern=s,x=samp.df$V1)]
}) %>%  unlist(.) %>% as.character(.) %>% unique(.) 

keep.vec <- unique(c(keep.afr,keep.eur,keep.amr))

build.df1 <- data.frame("V1"=keep.afr,"V2"="AFRICA",stringsAsFactors = F)
build.df2<- data.frame("V1"=keep.eur,"V2"="EUROPE",stringsAsFactors = F)
build.df3<- data.frame("V1"=keep.amr,"V2"="AMERICA",stringsAsFactors = F)

map.df <- rbind(build.df1,build.df2,build.df3)
```

```{r}
write.table(x=keep.vec,file=input.dir %&%"ref-samples.txt",sep="\t",
            quote=F,row.names=F,col.names=F)
write.table(x=map.df,
            file=input.dir %&%"ref-map.txt",sep="\t",
            quote=F,row.names=F,col.names=F)
```
